refactor(web-ui): split workflow view mode controller
Extract responsive Flow/Graph mode switching into a dedicated Stimulus controller so the editor controller focuses on graph data and rendering behavior. This keeps responsibilities clearer and makes future UI iteration safer.
